refactor(socket): extract Socket.IO setup and handlers out of server.js
server.js had grown to 830 lines, of which about 775 were Socket.IO
wiring. It is now a 50-line bootstrap: the Turnstile fail-closed guard,
the HTTP server, initSocket, the scheduled jobs, and listen.

The realtime code moves into src/socket/:

- index.js exposes initSocket(server, app), which creates the io
  instance with the existing CORS check, sets it on the app so
  controllers keep reaching it via req.app.get('io'), installs the auth
  middleware, and registers every handler on connection
- socketAuth.js holds the handshake authentication
- access.js holds the shared authorization helpers (team membership,
  direct conversation access, reply access, blocked user rooms, chat
  file expiry)
- presence.js holds the connected users map, room joins, users:online
  and disconnect
- handlers/ splits the event listeners by domain: conversation join and
  leave, message:new including the mention fanout, typing start and
  stop, and message:read

Handler bodies were moved verbatim so the diff stays reviewable; the
duplication between typing:start and typing:stop is left as is and can
be collapsed separately.

One deliberate change beyond the move: the lazy require of
notificationController inside message:new is now a top-level import.
notificationController only requires the database config, so there is
no cycle to work around.

Verified: node --check on all files, every module loads with its
exports resolved, a name-and-count diff of all socket .on and .emit
calls against the old file is identical, a normalized line diff shows
only scaffolding changes, 108/108 tests pass, the server boots with
Socket.IO enabled, and realtime chat was smoke tested in the browser.
